+++ src/OpenTelemetry.Resources.Host/HostSemanticConventions.cs
@@ -1,0 +1,10 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+namespace OpenTelemetry.Resources.Host;
+
+internal static class HostSemanticConventions
+{
+    public const string AttributeHostName = "host.name";
+    public const string AttributeHostId = "host.id";
+}